Rename client/token specs
