+++ e2e/specs/profile/clients.spec.js
@@ -1,5 +1,5 @@
-const { constants} = require('../../../constants');
-import { Profile, OauthCreateDrawer, dialogMap} from '../../../pageobjects/profile';
+const { constants} = require('../../constants');
+import { Profile, OauthCreateDrawer, dialogMap} from '../../pageobjects/profile';
 
 const profile = new Profile();
 const createDrawer = new OauthCreateDrawer();
